requirements.txt
# Bridge server (bridge/ package; run: python3 -m bridge or python3 bridge_server.py)
fastapi>=0.100.0
uvicorn>=0.22.0
httpx>=0.24.0
PyJWT>=2.0.0
websockets>=11.0
pydantic>=2.0.0
python-dotenv>=1.0.0
